Moving the Node.js tests to a new test folder
