+++ test/node/sth_app_test_configuration.js
